added folder option and prevented overwriting

+++ note.py
@@ -10,6 +10,10 @@ openai.api_key = os.getenv("OPENAI_API_KEY") # sets the API key
 title = input("Note Title: ")
 
 include_title = input("Include title in transcription? (y/n): ")
+include_folder = input("Include folder to save transcription inside? (y/n): ")
+
+if include_folder == "y":
+    folder = input("Folder name: ")
 
 # take in a stream of audio and save it as an audio file
 if __name__ == "__main__":
@@ -32,10 +36,28 @@ if include_title == "y":
 else:
   obsidian_note = f"{transcript}"
 
+if include_folder == "y":
+  obsidian_vault_path = f"{obsidian_vault_path}{folder}"
+  if not os.path.exists(obsidian_vault_path):
+    os.mkdir(obsidian_vault_path)
+  
 # Remove or replace invalid characters
 sanitized_title = re.sub(r'[\\/*?:"<>|]', '_', title)
 
-with open(f"{obsidian_vault_path}/{sanitized_title}.md", "w") as f:
-    f.write(obsidian_note)
+filepath = f"{obsidian_vault_path}/{sanitized_title}.md"
 
-print(f"Note saved as {sanitized_title}.md in Obsidian Vault.")
+if os.path.exists(filepath):
+  overwrite = input("File already exists. Overwrite? (y/n): ")
+  if overwrite == "y":
+    with open(f"{obsidian_vault_path}/{sanitized_title}.md", "w") as f:
+      f.write(obsidian_note)
+  else:
+    print("File not saved.")
+else:
+  with open(f"{obsidian_vault_path}/{sanitized_title}.md", "w") as f:
+      f.write(obsidian_note)
+
+if include_folder == "y":
+  print(f"Note saved as {sanitized_title}.md in Obsidian Vault\{folder}.")
+else:
+  print(f"Note saved as {sanitized_title}.md in Obsidian Vault.")
